restrict race detection to concurrent feed map write test
