+++ shared/p2p/BUILD.bazel
@@ -40,7 +40,6 @@ go_test(
         "topics_test.go",
     ],
     embed = [":go_default_library"],
-    race = "on",
     deps = [
         "//proto/sharding/v1:go_default_library",
         "//shared:go_default_library",
@@ -54,4 +53,13 @@ go_test(
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
+)
+
+go_test(
+    name = "go_feed_concurrent_write_test",
+    srcs = [
+        "feed_concurrent_test.go",
+    ],
+    embed = [":go_default_library"],
+    race = "on",
 )
